third_party/blink/web_tests/platform/mac/virtual/stylable-select-disabled/external/wpt/html/semantics/forms/the-select-element/stylable-select/select-keyboard-behavior.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] custombutton: When the listbox is closed, all arrow keys should open the listbox.
  assert_true: Arrow left should open the listbox. expected true got false
[FAIL] custombutton: When the listbox is closed, the enter key should submit the form or do nothing.
  assert_true: Enter should submit the form when the listbox is closed. expected true got false
[FAIL] custombutton: When the listbox is open, the enter key should commit the selected option.
  assert_equals: The selected option should receive initial focus. expected Element node <option class="two">two</option> but got Element node <select id="custombutton">\n    custom button\n    \n      <...
Harness: the test ran to completion.
